♻️ Extract account page state machines and modals
Move Stripe Connect and Liker Plus management logic into composables,
and the display-name/email/adult-content/delete-account dialogs into
overlay-created modal components.
